refactor: 페키지 경로 변경 및 RestAssuredFixture로 이름 변경
